Started work on a rope prototype
At the moment this is missing balancing.

A rope is a data structure that is able to handle concatenation and
insertion very efficiently, which makes it a good candidate for file
representation in a text editor. Unlike a gap buffer insertions remain
cheap even if the position changes randomly, this makes it specifically
interesting for collaborative text editing.

The downside of a rope is that the implementations and papers I've seen
so far assume that the entire string can be held in memory, which may
not always be practical.